Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.1
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.7.4
pydantic-settings==2.1.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.1.0 
httpx==0.28.1
requests==2.32.3
apscheduler==3.10.4
numpy==1.26.3
pandas==2.1.4
scipy==1.12.0
scikit-learn==1.4.0
statsmodels==0.14.4
plotly==5.18.0
kaleido==0.2.1
# AI Agent Dependencies
# These are additional dependencies required for the AI Agent functionality

# LangGraph for workflow orchestration
langgraph==0.4.7

# OpenAI for transcription and analysis
openai==1.82.0

# Computer Vision for frame extraction
opencv-python

# Additional dependencies that might be needed
typing-extensions